Implement a method that initializes game board
The method is defined in the game board module since it's an operation
that sets up the game board array.

In order for the method to be able to search for the dom elements, the
tic tac toe grid element was filled with 9 cell elements corresponding
to each cell in the tic tac toe grid.