Implement the game logic module
So far, the game logic module alternates the player turns whenever a
cell has been clicked on the tic tac toe grid.

The display controller module was also modified to make sure O's are
being displayed when it's player 1's turn and X's are being displayed
when it's player 2's. The display controller module also lets the user
know who's turn it is in a display above the tic tac toe grid.