Make sure the cell the user clicked is not taken
A function was created within the game logic module to check if the
cell that the user clicked on is occupied by an X or an O.

If this function returns true, turns will not be switched and a flag
will be switched to true to let the display controller module know not
to change anything in the display. An alert will also pop up to let the
user know to choose a different cell.