Add a new game button when a game ends
An event listener was added to the new game button in both the game
logic module and the display controller module.

The event listener within the game logic module makes sure the
variables that allow the game logic to function are reset to their
initial values.

The event listener within the display controller module makes sure the
text being displayed in the cells are cleared.